Improvements

+++ backend/fms_core/migrations/0019_v3_1_2.py
@@ -1,7 +1,6 @@
 from django.conf import settings
 from django.db import migrations
 from django.contrib.auth.models import User
-import django.utils.timezone
 import reversion
 import datetime
 
@@ -14,7 +13,7 @@ def fix_process_data(apps, schema_editor):
     ProcessSample = apps.get_model("fms_core", "ProcessSample")
     Process = apps.get_model("fms_core", "Process")
 
-    affected_revisions_ids = Revision.objects.filter(date_created__range=["2021-04-29", "2021-04-30"]).values_list('id', flat=True)
+    affected_revisions_ids = Revision.objects.filter(date_created__range=["2021-04-29", "2021-05-01"]).values_list('id', flat=True)
 
     with reversion.create_revision(manage_manually=True):
         reversion.set_comment("Curation of multiple processes into one process per template")
@@ -31,23 +30,24 @@ def fix_process_data(apps, schema_editor):
                 # Getting all Processes related to those ProcessSample
                 process_ids = list(process_samples.values_list('process_id', flat=True))
 
-                # First Process; the only one that will be kept
-                first_process_id = min(process_ids)
+                if len(process_ids) > 1:
+                    # First Process; the only one that will be kept
+                    first_process_id = min(process_ids)
 
-                # Change process_id for ProcessSample now pointing to the wrong processes
-                for ps in process_samples:
-                    ps.process_id = first_process_id
-                    ps.save()
-                    reversion.add_to_revision(ps)
+                    # Change process_id for ProcessSample now pointing to the wrong processes
+                    for ps in process_samples:
+                        ps.process_id = first_process_id
+                        ps.save()
+                        reversion.add_to_revision(ps)
 
-                # Delete Processes unused (should not have been created in first place)
-                Process.objects.filter(pk__in=process_ids).exclude(pk=first_process_id).delete()
+                    # Delete Processes unused (should not have been created in first place)
+                    Process.objects.filter(pk__in=process_ids).exclude(pk=first_process_id).delete()
 
-                # Update First Process for this revision batch with comment reflecting the current curation
-                first_process = Process.objects.get(id=first_process_id)
-                first_process.comment += f' - Curated {datetime.datetime.now().strftime("%Y-%m-%d")} v3.1.2 combined multiple processes'
-                first_process.save()
-                reversion.add_to_revision(first_process)
+                    # Update First Process for this revision batch with comment reflecting the current curation
+                    first_process = Process.objects.get(id=first_process_id)
+                    first_process.comment += f' - Curated {datetime.datetime.now().strftime("%Y-%m-%d")} v3.1.2 combined multiple processes'
+                    first_process.save()
+                    reversion.add_to_revision(first_process)
 
 
 class Migration(migrations.Migration):
